requirements.txt
streamlit
pytesseract
opencv-python
numpy
Pillow
easyocr
tensorflow
reportlab
pdf2image
google-cloud-vision
